extra embedded resource pass
Added a loop between plugin assembly loading and instantiation that resolves embedded DLLs.
Modified `AssemblyUtils::GetMissingDependencies` to create a set of the loaded assemblies.
Changed the `LoadEmbeddedDll` and `LoadCompressedEmbeddedDll` methods in `AssemblyUtils` class to write resolved the assembly to a temporary file, load the DLL if no assembly with the same name has been loaded. The file is deleted after this operation.