IsAbstract plugin check + separate embedded dependency resolving pass (#212)
* don't instantiate abstract plugins

Added a condition to skip abstract classes when instantiating plugins.

* extra embedded resource pass

Added a loop between plugin assembly loading and instantiation that resolves embedded DLLs.
Modified `AssemblyUtils::GetMissingDependencies` to create a set of the loaded assemblies.
Changed the `LoadEmbeddedDll` and `LoadCompressedEmbeddedDll` methods in `AssemblyUtils` class to write resolved the assembly to a temporary file, load the DLL if no assembly with the same name has been loaded. The file is deleted after this operation.

* rearrange & slightly optimize code

Moved private methods in AssemblyUtils & PluginLoader to the end of the file.
Made the LINQ selectors in `LogMissingDependencies` and `GetLoadedAssemblies` static.
Removed the `ToArray` call and made the variable an `IEnumerable<string>` in `LogMissingDependencies` to improve performance.

* Apply suggestions from code review

Co-authored-by: Jesus QC <69375249+Jesus-QC@users.noreply.github.com>

* add embedded dlls to Dependencies set

---------

Co-authored-by: Jesus QC <69375249+Jesus-QC@users.noreply.github.com>